# server/openapi.yaml
openapi: 3.0.3
info:
  title: LEGION
  version: 0.1.0
  description: The engine. One Django server, every limb a client of it.
paths:
...
  /api/body/bodyweight_logs/:
    get:
      operationId: api_body_bodyweight_logs_list
      description: |-
        `?since=<iso>` (tombstones included) and `?active=1` (live rows
        only). They compose: `?active=1&since=<iso>` is a live-rows-changed
        feed, and `?active=1` alone is every live row since the epoch.

        A missing `since` means EVERYTHING, never nothing - `api/sync.parse_since`
        holds that rule for the whole API and quotes the phone-side cursor
        comment it comes from.
      parameters:
      - in: query
        name: active
        schema:
          type: string
          enum:
          - '1'
          - 'on'
          - 'true'
          - 'yes'
        description: Narrows the feed to live rows (`deleted_at is null`). Composes
          with `since`. Anything else, `0` and `false` included, means the narrowing
          was not asked for and tombstones stay in - a caller who cannot be understood
          sees too much, never silently nothing.
      - in: query
        name: since
        schema:
          type: string
          format: date-time
        description: ISO-8601 UTC watermark. Returns rows with `updated_at >= since`,
          tombstones included, oldest first. **Omitted or unparsable means EVERYTHING,
          never nothing** (api/sync.parse_since), so a fresh client's first pull is
          the whole table. Hand back the `next` cursor from the previous page, or
          `server_time` from GET /api/changes; both are already `Z`-suffixed so they
          need no extra encoding.
      tags:
      - body
      security:
      - deviceToken: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedBodyweightLog'
          description: ''
  /api/body/bodyweight_logs/{identity}/:
    put:
      operationId: api_body_bodyweight_logs_update
      description: |-
        `PUT <table>/<identity>/`. Idempotent by construction: the
        identity comes from the URL, so a retry cannot make a second row.
        200 whether the row was created or updated - the caller asked for
        the row to exist in this state and it does, and which of the two
        happened is not something a retrying client can act on.
      parameters:
      - in: path
        name: identity
        schema:
          type: string
        description: The row's `origin_guid`.
        required: true
      tags:
      - body
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyweightLog'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BodyweightLog'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BodyweightLog'
        required: true
      security:
      - deviceToken: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BodyweightLog'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: 'The write was refused and NOTHING was written. Two body shapes
            occur and both are JSON objects: `{"detail": "..."}` for an unknown field,
            a database refusal or a hand-written check, and `{"<field>": ["..."]}`
            for a field-level validation error. Both carry text meant to be shown
            to a person.'
    delete:
      operationId: api_body_bodyweight_logs_destroy
      description: |-
        `DELETE <table>/<identity>/`. Sets the tombstone column; the row
        stays, because a phone that has not synced since still needs to
        learn the row is gone. Idempotent - a second delete is still a 204,
        matching `EventsBackend.softDelete`'s own contract.
      parameters:
      - in: path
        name: identity
        schema:
          type: string
        description: The row's `origin_guid`.
        required: true
      tags:
      - body
      security:
      - deviceToken: []
      responses:
        '204':
          description: Done. Soft-deleted (the tombstone row stays so an unsynced
            client learns of it).
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: No such row. Nothing was changed.
...
components:
  schemas:
    BodyweightLog:
      type: object
      description: |-
        Base for every serializer this viewset drives.

        Two behaviours, both ticket 04's own text:

        1. **Unknown fields are refused, never dropped.** "A limb sending a
           column the server does not know is a version skew, and it says so
           with a 400 naming the field, never a silent drop." A phone one
           release ahead, writing a column this server has never heard of,
           must be told - a silent drop would look exactly like a successful
           write and lose the value.
        2. **The server facts are the server's.** `id`, `provenance`,
           `created_at`, `updated_at` and `deleted_at` are read-only on every
           subclass; `deleted_at` moves only via DELETE, `updated_at` only via
           the `touch_updated_at` trigger.

        Subclasses set `Meta.model` / `Meta.fields` and, where the table's
        `provenance` column is not the shared `USER` default,
        `default_provenance`.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        weight_value:
          type: number
          format: double
        weight_unit:
          type: string
        logged_at:
          type: string
          format: date-time
        trust_tier:
          type: string
        provenance:
          allOf:
          - $ref: '#/components/schemas/ProvenanceEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        origin_guid:
          type: string
      required:
      - created_at
      - deleted_at
      - id
      - logged_at
      - origin_guid
      - provenance
      - trust_tier
      - updated_at
      - weight_unit
      - weight_value
...
    Detail:
      type: object
      description: |-
        `{"detail": "..."}` - the shape every hand-written refusal in this API
        uses, and the shape DRF itself uses for 401/403/404/405. The strings are
        written to be shown to a person: they say what did NOT happen (CLAUDE.md
        section 7's outcome-verb rule), so a client should surface `detail`
        rather than inventing its own wording.
      properties:
        detail:
          type: string
      required:
      - detail
...
    PagedBodyweightLog:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/BodyweightLog'
        next:
          type: string
          format: date-time
          nullable: true
          description: 'Cursor for the next page: hand it back as `?since=`. Null
            means this was the last page.'
      required:
      - next
      - results
...
    ProvenanceEnum:
      enum:
      - DETERMINISTIC
      - LLM_RECONCILED
      - UNRECONCILED
      - USER
      type: string
      description: |-
        * `DETERMINISTIC` - Deterministic
        * `LLM_RECONCILED` - Llm Reconciled
        * `UNRECONCILED` - Unreconciled
        * `USER` - User
...
  securitySchemes:
    deviceToken:
      type: apiKey
      in: header
      name: Authorization
      description: One token per device (phone, browser, future robot), revocable
        alone. Send the value as `Token <key>`, including the word Token and the space.
        Get a key from POST /api/auth/login; POST /api/auth/logout revokes only the
        key that made the request.